Update Mockup_Themes/COS-dk/js/cos.js
Noticed that "console.log( top_level.children('ul').length );" was keeping IE from correctly displaying the custom sidebar nav.  I've replaced it with a trimmed down version.  If you end up on a page that is not part of the Custom Menu it will no longer say "null" it will hide the menu entirely.