Massive Stying Update
-category.php - Commented out the get_search_form, removed the sidebar
div, and added a new class "fullwidth" to make this page span the full
width.
-hope.less - Reduced the width of #third-footer-widget-area down to
235px.
-layout.less - Reduced the #container min-hieght to 100px, and removed
the margin-bottom, removed some commented out code, changed the main >a
link color to #08C to match the Comm site.  Adjusted the margin top for
the home page #left-content and #right-content.  Removed the
innerContent margin bottom and adjusted the td verticle-align to top for
the calendar plugin.  Added the class fullwidth .column(12).  Adjusted
the sidebar's margins.  Gave textwidget a line-height.  Adjusted the
#main_footer margins and section padding.  Changed the h2 for the news
section to be the correct font family, weight, color, size, line-height,
etc.  Added 3 classes for the first, second and third footer widget
areas, as well as added new classes for .news and #widget_news.
Adjusted the #inner_searchform and #searchform styling to match the Comm
site.
-functions.php - Manually changed the news feed option to be hard coded
to the Comm site.  Added in the new class "widget_news" to the show
shows function while adjusting the formating to match the news on the
Comm site.  Removed the registration of the Front Page Slider Right.
-header.php - Added in the favicon from the Comm site
-loop-newscat.php - Commented out the Search from the top of the page
and removed the sidebar
-onecolumn-page.php - Changed the get_template_part to call loop
fullwidth.
-search.php - Added the new class fullwidth
-searchform.php - Added in a button form the Comm site to make it match
-single.php - Changed the get_template_part to call loop fullwidth
-loop-fullwidth.php - Created new template page for fullwidth pages.