Better Compliance and Practices
=Updated the header.php file by removing the <link> to jquery and moved
the <link> for flexslider jquery into functions.php to be registered and
enqueued.
=Updated functions.php to wp_enqueue jquery since it's included with
wordpress.  This is a better way than a traditional <link> because it
will only load if it hasn't been loaded.  As previously mentioned, I
used a wp_registere_script & wp_enqueue_script on the flexslider js
inside of a is_home call so that it only happens on the homepage.
=Updated cos.js because by using wp_enqueue_script for jquery we're
telling WordPress to load jquery in "no conflict" mode.   Changed
"$(function() {" to "jQuery(document).ready(function ($) {" so that the
$ shortcut can continue to be used in the file.