General Updates
/**** CSS Updates ****/
Accommodate when there are no items in the footer.
Remove >> from lists in the "All in 1 Event Calendar"
/**** Functions ****/
Corrected the order of the enqueuing of scripts so that it properly
enabled jQuery.  The cos.js enqueue cannot come before the wp_enqueue
for jQuery since it relies on it.
/**** Header ****/
Updated a few items to be more w3c compliant
/**** Misc Files ****/
Uploaded the different custom post type photos and login banner image
for COS